Improved whois command

+++ main.py
@@ -25,12 +25,14 @@ async def on_command_error(ctx, error):
     elif isinstance(error, discord.ext.commands.errors.MissingPermissions) or isinstance(error, discord.Forbidden):
         await ctx.send("Sorry. You don't have the permission for that command.")
 
-    elif isinstance(error, discord.ext.commands.errors.CommandOnCooldown):
+    elif isinstance(error, discord.ext.commands.errors.CommandOnCooldown) and ctx.command.name == "report":
         await ctx.message.delete()
-        await ctx.author.send(f"""Sorry, but we have made a cooldown to prevent spamming. Try again in {error.retry_after:,.2f} seconds.
+        await ctx.author.send(f"""Sorry, but we have made a cooldown to prevent the abuse of the command. Try again in {error.retry_after:,.2f} seconds.
 If you want to report something before the cooldown is over or you made a report on accident then please contact a staff member and we will get it sorted out.""")
+    elif isinstance(error, discord.ext.commands.errors.CommandOnCooldown):
+        await ctx.send(f"You need to wait {error.retry_after:,.2f} seconds before trying this command again.")
 
-    else: await ctx.send(error)
+    else: ctx.send(error)
 
 
 for filename in os.listdir('./cogs'):
